Fix compiler and code-model warnings

+++ src/monitor-lib/sysfsreader.cpp
@@ -44,7 +44,7 @@
 #include <qplatformdefs.h>
 #include "sysfsreader.h"
 
-#  define EINTR_LOOP(cmd) __extension__ ({__typeof__(cmd) res = 0; do { res = cmd; } while (res == -1 && errno == EINTR); res; })
+#define EINTR_LOOP(cmd) __extension__ ({__typeof__(cmd) res = 0; do { res = cmd; } while (res == -1 && errno == EINTR); res; })
 
 static inline int qt_safe_open(const char *pathname, int flags, mode_t mode = 0777)
 {
@@ -56,8 +56,9 @@ static inline int qt_safe_open(const char *pathname, int flags, mode_t mode = 07
         ::fcntl(fd, F_SETFD, FD_CLOEXEC);
     return fd;
 }
-#  undef QT_OPEN
-#  define QT_OPEN         qt_safe_open
+
+#undef QT_OPEN
+#define QT_OPEN         qt_safe_open
 
 QT_BEGIN_NAMESPACE_AM
 
@@ -94,7 +95,7 @@ QByteArray SysFsReader::readValue() const
     int offset = 0;
     int read = 0;
     do {
-        read = static_cast<int>(EINTR_LOOP(QT_READ(m_fd, m_buffer.data() + offset, m_buffer.size() - offset)));
+        read = int(EINTR_LOOP(QT_READ(m_fd, m_buffer.data() + offset, size_t(m_buffer.size() - offset))));
         if (read < 0)
             return QByteArray();
         else if (read < (m_buffer.size() - offset))
